Declare 'svunit' as a source dependency
We would like to use the 'svunit-build' plugin to build it, but it's
currently not possible to inject custom plugins. To work around this
limitation I tried to wrap the custom plugin in a "local" plugin from
an included build. This doesn't work, though, because when running we
get an error that Gradle can't resolve the 'svunit-build' plugin.