Add workaround to custom plugin apply in injected build
Gradle seems to have issues with setting the correct runtime classpath
in the injected build.

Also moved to a newer version of 'gradle-hdvl'. The 'svunit-build' is
now a 'Project' plugin. We can only do this hack for 'Project' plugins,
because the class loader for the 'Settings' object is locked and we
can't fiddle with it.